agent/app/requirements.txt
fastapi==0.115.0
langchain==0.1.10
openai==1.13.3
python-dotenv==1.0.1
uvicorn==0.31.0
langchain-core==0.1.30
langchain-openai==0.0.8